+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/16 14:50:11 by aben-ham          #+#    #+#             */
-/*   Updated: 2021/11/20 19:08:38 by aben-ham         ###   ########.fr       */
+/*   Updated: 2021/11/22 17:33:53 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 // strlen count the number of characters in a given string
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != 0)
